examples/front_ends/microsoft_agents_365/simple_web_query/requirements.txt
--pre

aioboto3==15.5.0
aiobotocore==2.25.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==26.1.0
Authlib==1.6.9
azure-core==1.39.0
beautifulsoup4==4.14.3
blinker==1.9.0
boto3==1.40.61
botocore==1.40.61
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
cryptography==46.0.5
dataclasses-json==0.6.7
datasets==4.8.4
dill==0.4.1
distro==1.9.0
et_xmlfile==2.0.0
expandvars==1.1.2
fastapi==0.135.2
fastuuid==0.14.0
filelock==3.25.2
filetype==1.2.0
Flask==3.1.3
frozenlist==1.8.0
fsspec==2026.2.0
googleapis-common-protos==1.73.0
grpcio==1.80.0rc1
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.1
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
jsonpatch==1.33
jsonpath-ng==1.8.0
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.13
langchain-aws==1.1.0
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.21
langchain-huggingface==1.2.1
langchain-litellm==0.6.1
langchain-milvus==0.3.3
langchain-nvidia-ai-endpoints==1.2.1
langchain-openai==1.1.12
langchain-tavily==0.2.17
langchain-text-splitters==1.1.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
litellm==1.82.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
microsoft-agents-a365-notifications==0.2.1.dev44
microsoft-agents-a365-observability-core==0.2.1.dev44
microsoft-agents-a365-observability-hosting==0.2.1.dev44
microsoft-agents-a365-runtime==0.2.1.dev44
microsoft-agents-a365-tooling==0.2.1.dev44
microsoft-agents-activity==0.9.0.dev2
microsoft-agents-authentication-msal==0.9.0.dev2
microsoft-agents-hosting-aiohttp==0.9.0.dev2
microsoft-agents-hosting-core==0.9.0.dev2
msal==1.35.1
multidict==6.7.1
multiprocess==0.70.19
mypy_extensions==1.1.0
narwhals==2.18.0
nest-asyncio2==1.7.2
networkx==3.6.1
numpy==2.4.3
nvidia-nat==1.6.0b1
nvidia-nat-core==1.6.0b1
nvidia-nat-eval==1.6.0b1
nvidia-nat-langchain==1.6.0b1
nvidia-nat-opentelemetry==1.6.0b1
openai==2.29.0
openevals==0.1.3
openinference-semantic-conventions==0.1.28
openpyxl==3.1.5
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
optuna==4.8.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
pkce==1.0.3
pkginfo==1.12.1.2
platformdirs==4.9.4
plotly==6.6.0
propcache==0.4.1
protobuf==6.33.6
pyarrow==23.0.1
pycparser==3.0
pydantic==2.13.0b2
pydantic-settings==2.13.1
pydantic_core==2.42.0
Pygments==2.19.2
PyJWT==2.12.1
pymilvus==2.6.10
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
s3transfer==0.14.0
simsimd==6.5.16
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.1.0b1
starlette==0.52.1
tabulate==0.10.0
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
usearch==2.23.0
uuid_utils==0.14.1
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.7
wikipedia==1.4.0
wrapt==1.17.4rc1
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
microsoft-agents-a365-observability-extensions-langchain==0.2.1.dev48
